settings: show log file location
